\keyword{thin}
\name{vegman}
\alias{vegman}

\title{Allows user to thin vegetation to a target percent cover} 

\description{Thin vegetation to a target percent cover.}

\details{
	The \code{vegman} function has two arguments.  The first argument is a sample.data.data object.  The second argument is the 
	desired percentage of vegetation cover for the sample after treatment.
}

\author{Nathaniel L. Osborne \email{nathaniel.osborne@oregonstate.edu}}

\seealso{
  \code{\link{calc.max.sdi}},
  \code{\link{impute}},
  \code{\link{plants.cips}},
  \code{\link{plants.smc}},
  \code{\link{plants.swo}},
  \code{\link{plants.swohybrid}},
  \code{\link{plot.sample.data}},
  \code{\link{plots.cips}},
  \code{\link{plots.smc}},
  \code{\link{plots.swo}},
  \code{\link{plots.swohybrid}},
  \code{\link{print.sample.data}},
  \code{\link{project}},
  \code{\link{rand.seed}},
  \code{\link{rconifers}},
  \code{\link{sample.data}},
  \code{\link{set.species.map}},
  \code{\link{set.variant}},
  \code{\link{species.cips}},
  \code{\link{species.smc}},
  \code{\link{species.swo}},
  \code{\link{species.swohybrid}},
  \code{\link{sp.sums}},
  \code{\link{summary.sample.data}},
  \code{\link{thin}},
  \code{\link{variants}}
}

\keyword{treatment}


